requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-multipart
phonenumbers
python-dotenv
twilio
httpx
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dateutil
sqlmodel
sqlalchemy
alembic
python-jose[cryptography]
